fix: better error handling when pulling blog posts
As far as I can tell, errors /api/blog are causing 500s when Google tries to crawl the homepage. Try to do a little error handling in an attempt to prevent this.

I'm a little worried because behavior locally was very strange - I tried to get the server API to simply error but I saw the page rendering a non-error. This may be caching, in which case I hope it goes away in staging and prod.